requirements.txt
fastapi==0.115.12
uvicorn[standard]==0.34.2
celery==5.4.0
redis==5.0.7
sqlalchemy==2.0.38
psycopg[binary]==3.2.6
pydantic==2.11.3
pydantic-settings==2.8.1
python-multipart==0.0.20
playwright==1.52.0
beautifulsoup4==4.13.3
tldextract==5.1.2
python-whois==0.9.5
xgboost==2.1.4
joblib==1.4.2
openai==1.76.0
minio==7.2.15
httpx==0.28.1
docker==7.1.0
